debugger/requirements.txt
anthropic
langgraph
langchain-anthropic
python-dotenv
rich